feat:moved content of the header to Header component
