+++ src/components/Header.js
@@ -1,0 +1,12 @@
+import Navbar from "./Navbar";
+
+const Header = () => {
+    return (
+        <header>
+            <h1><u>My Library Helper</u></h1>
+            <Navbar />
+      </header>
+    )
+}
+
+export default Header;
